Refactor : pr 코멘트 반영
